+++ src/com/nixinova/graphics/BlocksRenderer.java
@@ -2,13 +2,12 @@ package com.nixinova.graphics;
 
 import com.nixinova.Conversion;
 import com.nixinova.coords.BlockCoord;
+import com.nixinova.coords.Coord;
 import com.nixinova.coords.PxCoord;
-import com.nixinova.coords.SubBlockCoord;
+import com.nixinova.coords.TxCoord;
 import com.nixinova.main.Game;
-import com.nixinova.player.Player;
 
 public class BlocksRenderer extends Render {
-
 	private Game game;
 
 	public BlocksRenderer(Render render, Game game) {
@@ -17,49 +16,79 @@ public class BlocksRenderer extends Render {
 		super.clearImage();
 	}
 
-	public void renderWorld() {
+	public void renderWorld(Game game) {
 		// Loop from bottom to top of world
 		BlockCoord min = this.game.world.minCorner;
 		BlockCoord max = this.game.world.maxCorner;
-		for (int y = min.y; y <= max.y; y++) {
-			for (int x = min.x; x <= max.x; x++) {
+		for (int x = min.x; x <= max.x; x++) {
+			for (int y = min.y; y <= max.y; y++) {
 				for (int z = min.z; z <= max.z; z++) {
-					// if (this.game.world.isExposed(x, y, z))
+					// Ensure block is not air
+					if (this.game.world.isAir(x, y, z))
+						continue;
+
+					// Render block if visible
+					boolean isVisible = Raycast.isBlockVisibleToPlayer(this.game, x, y, z);
+					if (isVisible) {
 						this.renderOneBlock(x, y, z);
+					}
+				}
+			}
+		}
+
+	}
+
+	private void renderOneBlock(int blockX, int blockY, int blockZ) {
+		final int size = Conversion.PX_PER_BLOCK;
+
+		// Loop through texels and render
+		// TODO fill inbetween as well
+		TxCoord startTx = Coord.fromBlock(blockX, blockY, blockZ).toTx();
+		for (int txX = 0; txX < size; txX++) {
+			for (int txY = 0; txY < size; txY++) {
+				for (int txZ = 0; txZ < size; txZ++) {
+					// Only render outside faces of the block;
+					boolean atStart = txX == 0 || txY == 0 || txZ == 0;
+					boolean atEnd = txX == size - 1 || txY == size - 1 || txZ == size - 1;
+					if (!atStart && !atEnd)
+						continue;
+
+					// Render this texel
+					this.renderOneTx(startTx.x + txX, startTx.y + txY, startTx.z + txZ);
 				}
 			}
 		}
 	}
 
-	public void renderOneBlock(int blockX, int blockY, int blockZ) {
-		Render texture = this.game.world.getTextureAt(blockX, blockY, blockZ);
+	private void renderOneTx(int txX, int txY, int txZ) {
+		BlockCoord blockCoord = Coord.fromTx(txX, txY, txZ).toBlock();
+		Render texture = this.game.world.getTextureAt(blockCoord.x, blockCoord.y, blockCoord.z);
 
 		// Skip if air
 		if (texture == null) {
 			return;
 		}
 
-		PxCoord posOnScreen = blockCoordToScreenPx(blockX, blockY, blockZ);
+		PxCoord posOnScreen = txCoordToScreenPx(txX, txY, txZ);
 		if (posOnScreen != null) {
-			// Render block
-			BlockCoord blockPos = new BlockCoord(blockX, blockY, blockZ);
-			this.generateRenderedBlock(texture, blockPos, posOnScreen);
+			// Render texel
+			int txPixel = Texture.getTexel(texture, txX, txZ);
+			this.generateRenderedTexel(txPixel, posOnScreen, blockCoord);
 		}
 	}
 
-	public PxCoord blockCoordToScreenPx(int blockX, int blockY, int blockZ) {
-		SubBlockCoord playerPos = this.game.controls.getPosition().toSubBlock();
-		playerPos.y += Player.PLAYER_HEIGHT;
+	private PxCoord txCoordToScreenPx(int txX, int txY, int txZ) {
+		PxCoord camPos = this.game.controls.getCameraPosition().toPx();
 		double rot = this.game.controls.getXRot();
 		double tilt = this.game.controls.getYRot();
 
 		// Relative position of block in world and player pos
-		SubBlockCoord relPos = new SubBlockCoord(blockX - playerPos.x, blockY - playerPos.y, blockZ - playerPos.z);
+		PxCoord relPos = new PxCoord(txX - camPos.x, txY - camPos.y, txZ - camPos.z);
 		double absDistance = Math.sqrt(relPos.x * relPos.x + relPos.y * relPos.y + relPos.z * relPos.z);
 
 		// Apply Y-axis (horiz) rotation
 		double xRot = relPos.x * Math.cos(rot) - relPos.z * Math.sin(rot);
-		double yRot = relPos.y;// * Math.cos(tilt) - relPos.z * Math.sin(tilt);
+		double yRot = relPos.y;
 		double zRot = relPos.x * Math.sin(rot) + relPos.z * Math.cos(rot);
 
 		// Apply X-axis (vertical) tilt
@@ -79,40 +108,24 @@ public class BlocksRenderer extends Render {
 		return new PxCoord(screenX, screenY, absDistance);
 	}
 
-	public void generateRenderedBlock(Render texture, BlockCoord blockPos, PxCoord screenPos) {
-		final int TEX_SIZE = Conversion.PX_PER_BLOCK;
-		final int baseBlockSize = 12; // temporary measure
+	private void generateRenderedTexel(int pixel, PxCoord screenPos, BlockCoord blockPos) {
+		final int TEXTURE_SIZE = Conversion.PX_PER_BLOCK;
+		final int TEXEL_SIZE = 16;
 
 		int startX = (int) screenPos.x;
 		int startY = (int) screenPos.y;
-		double relDepth = screenPos.z;
 
-		int size = (int) (TEX_SIZE * TEX_SIZE * baseBlockSize / relDepth);
-
-		for (int x = 0; x < size; x++) {
-			for (int y = 0; y < size; y++) {
-				int screenX = startX + x - size / 2;
-				int screenY = startY + y - size / 2;
+		for (int x = 0; x < TEXEL_SIZE; x++) {
+			for (int y = 0; y < TEXEL_SIZE; y++) {
+				int screenX = startX + x - TEXTURE_SIZE;
+				int screenY = startY + y - TEXTURE_SIZE;
 
 				// Ensure pixel is within screen bounds
 				if (isValidPosition(screenX, screenY)) {
-					// Set textured screen pixel
-					int texelX = x * TEX_SIZE / size;
-					int texelY = y * TEX_SIZE / size;
-					setPixel(screenX, screenY, Texture.getTexel(texture, texelX, texelY));
-
-					// Save selected block to player if in center of screen
-					if (getPixelIndex(screenX, screenY) == imageSize() / 2) {
-						setAsSelectedBlock(blockPos);
-					}
-
+					setPixel(screenX, screenY, pixel);
 				}
 			}
 		}
-	}
-
-	private void setAsSelectedBlock(BlockCoord blockPos) {
-		this.game.player.setLookingAt(blockPos.x, blockPos.y, blockPos.z);
 	}
 
 }