+++ src/com/nixinova/graphics/ScreenText.java
@@ -1,8 +1,10 @@
 package com.nixinova.graphics;
 
 import java.awt.Color;
+import java.awt.Font;
 import java.awt.Graphics;
 
+import com.nixinova.coords.Coord;
 import com.nixinova.coords.SubBlockCoord;
 import com.nixinova.main.Game;
 import com.nixinova.main.Mineo;
@@ -20,15 +22,39 @@ public class ScreenText {
 	}
 
 	public void drawMainInfo(Game game) {
-		SubBlockCoord playerBlockPos = game.player.position.toSubBlock();
-		String playerX = String.format("%.1f", playerBlockPos.x);
-		String playerY = String.format("%.1f", playerBlockPos.y);
-		String playerZ = String.format("%.1f", playerBlockPos.z);
+		int i = 0;
+		
+		Font initialFont = graphics.getFont();
 
+		float headingFontSize = 1.25f;
+		Font headingFont = initialFont.deriveFont(initialFont.getSize() * headingFontSize);
+		graphics.setFont(headingFont);
 		graphics.setColor(Color.white);
-		graphics.drawString(Mineo.TITLE, 5, SEP);
-		graphics.drawString("FPS: " + String.valueOf(game.fps), 5, SEP * 2);
-		graphics.drawString("Block: " + playerX + " / " + playerY + " / " + playerZ, 5, SEP * 3);
+		graphics.drawString(Mineo.TITLE, INDENT, SEP * ++i);
+
+		Coord position = game.player.getPosition();
+		graphics.setFont(initialFont);
+		
+		// FPS
+		var fpsFmt = String.valueOf(game.fps);
+		graphics.drawString("FPS: " + fpsFmt, INDENT, SEP * ++i);
+		// Player block at foot
+		graphics.drawString("Block: " + formatCoord(position), INDENT, SEP * ++i);
+		// Lowest ground Y value
+		var blockPos = position.toBlock();
+		var groundY = game.world.getGroundY(blockPos.x, blockPos.z);
+		graphics.drawString("Lowest ground: " + groundY, INDENT, SEP * ++i);
+		// Camera pos and angle
+		var camPosFmt = formatCoord(game.controls.getCameraPosition());
+		graphics.drawString("Camera: " + camPosFmt, INDENT, SEP * ++i);
+		// Mouse look angle
+		var mouseLookFmt = String.format("%.1f / %.1f",
+			(Math.toDegrees(game.controls.getXRot()) + 360) % 360, Math.toDegrees(game.controls.getYRot()));
+		graphics.drawString("Angle: " + mouseLookFmt, INDENT, SEP * ++i);
+		// Block being looked at
+		var lookingAt = game.player.getLookingAt();
+		var lookingAtFmt = lookingAt == null ? "none" : (lookingAt.x + " / " + lookingAt.y + " / " + lookingAt.z);
+		graphics.drawString("Hovered: " + lookingAtFmt, INDENT, SEP * ++i);
 	}
 
 	public void drawOptionsWarning() {
@@ -55,6 +81,14 @@ public class ScreenText {
 		graphics.drawString(msg1, INDENT, Display.HEIGHT - SEP * 8);
 		graphics.drawString(msg2, INDENT, Display.HEIGHT - SEP * 7);
 		graphics.drawString(msg3, INDENT, Display.HEIGHT - SEP * 6);
+	}
+
+	private String formatCoord(Coord coord) {
+		SubBlockCoord pos = coord.toSubBlock();
+		String playerX = String.format("%.1f", pos.x);
+		String playerY = String.format("%.1f", pos.y);
+		String playerZ = String.format("%.1f", pos.z);
+		return playerX + " / " + playerY + " / " + playerZ;
 	}
 
 }